refactor(Package): domain 패키지 내부 car 패키지 추가
NumberGenerator, Car, ParticipateCars car 패키지로 이동
