+++ src/main/java/racingcar/RacingCarApplication.java
@@ -3,7 +3,7 @@ package racingcar;
 import racingcar.domain.GameInfo;
 import racingcar.domain.GameResult;
 import racingcar.domain.RacingGame;
-import racingcar.domain.common.RandomNumberGenerator;
+import racingcar.domain.car.RandomNumberGenerator;
 import racingcar.dto.InputValueDto;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
