+++ src/main/java/racingcar/domain/car/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain.common;
+package racingcar.domain.car;
 
 public interface NumberGenerator {
     int generateNumber();
