+++ src/test/java/racingcar/domain/car/CarTest.java
@@ -1,9 +1,9 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.domain.car.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
