+++ src/test/java/racingcar/domain/car/FixedNumberGenerator.java
@@ -1,4 +1,6 @@
-package racingcar.domain.common;
+package racingcar.domain.car;
+
+import racingcar.domain.car.NumberGenerator;
 
 public class FixedNumberGenerator implements NumberGenerator {
     private static final int MOVE_CONDITION_NUMBER = 4;
