+++ src/test/java/racingcar/domain/car/ParticipateCarsTest.java
@@ -1,15 +1,14 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.domain.common.FixedNumberGenerator;
+import racingcar.domain.car.FixedNumberGenerator;
+import racingcar.domain.car.ParticipateCars;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Random;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
